app/requirements.txt
uvicorn==0.23.1
fastapi==0.100.0
python-dotenv==1.0.0
pydantic==2.10.6
requests==2.32.2
python-binance==1.0.28
pandas==2.2.3
